+++ CHANGELOG.md
@@ -1,4 +1,27 @@
+## 0.0.7
+
+- Upgrade IronSource SDK Version
+- Fixed Admob Bug
+
+## 0.0.6
+
+- Fixed TO DO Bug
+
+## 0.0.5
+
+- Fixed init Bug
+- Banner Size Type
+- Update manual for Banner Size Type
+
+## 0.0.4
+
+- Fixed NullSafety Bug
+- Upgrade IronSource SDK Version
+- Upgrade Flutter SDK Version
+- Fixed Banner Bugs in example
+
 ## 0.0.3
+
 - Update IronSource SDK Version
 - Fixed Banner Bugs in example